Refactored drawer implementation
